Animation!!!
The whole point of the animation library was to be better and faster than doing it manually. SO much time wasted for it not to work.

Surprise! Doing it manually was super effective.

The BFG projectile now travels and explodes after a set time. I plan to add collision to it next.